fix(DeleteBadge): add error handling for when badge with provided id does not exist

+++ ferrilata-devilline/Controllers/ApiControllers/BadgesController.cs
@@ -31,8 +31,11 @@ namespace ferrilata_devilline.Controllers.ApiControllers
         [Route("/api/badges/{badgeId}")]
         public IActionResult DeleteBadge(long badgeId)
         {
+            if (_badgeService.FindBadgeById(badgeId) == null)
+                return BadRequest(new {error = "Requested Badge does not exist"});
             _badgeService.DeleteById(badgeId);
             return Ok("Deleted");
+
         }
 
         [HttpDelete]
